Fix input file in README test commands
As currently presented in the README, the commands for testing Singularity containers use an input file that still has the GMT/GhostScript paths set. This means that people who try to run the commands in the README will get confusing errors. Changing the input file in the sample commands fixes that.